requirements.txt
google-auth==1.14.3
google-auth-oauthlib==0.4.1
google-pasta==0.1.7
googleapis-common-protos==1.51.0
google-api-python-client
httplib2==0.9.2
numpy==1.16.5
oauth==1.0.1
oauthlib==3.1.0
requests==2.22.0
requests-oauthlib==1.3.0
simplejson==3.13.2
testpath==0.4.4
urllib3==1.25.8
virtualenv==15.1.0
